requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
elasticsearch==5.4.0
enum34==1.1.6
fastdtw==0.3.0
Flask==0.12.2
functools32==3.2.3.post2
get==0.0.0
gunicorn==19.7.1
html==1.16
idna==2.6
implicit==0.2.7
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.0.2
mogrify==0.1
mysql-connector==2.1.4
nbformat==4.4.0
numpy==1.13.3
pandas==0.21.0
plotly==2.2.1
post==0.0.0
public==0.0.0
pyechonest==9.0.0
pyparsing==2.2.0
pyrecsys==0.0.3
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
query-string==0.0.0
recommend==0.2.1
recsys==0.0.4
redis==2.10.6
request==0.0.0
requests==2.18.4
scikit-learn==0.18.1
scikit-surprise==1.0.4
scipy==0.19.0
seaborn==0.8.1
setupfiles==0.0.50
six==1.10.0
sklearn==0.0
subprocess32==3.2.7
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
